Add an xfail py_test to verify rust panics in cc
For some reason, EXPECT_ANY_THROW in gtest doesn't catch
panics on rust side and the test executable crashes. EXPECT_DEATH
wasn't working either, so I decided to wrap it in a py_test
that depends on a separate small cc_test executable.

The py_test captures stderr, verifies rust panic strings are
present and return code is non-zero. The rust panic strings
rely on the RUST_BACKTRACE=full env settings, which is also
in the build system